renamed string.h to KarplusString.h
