+++ Source/PhysicalModeling/KarplusString.h
